Fix a division by zero bug
This fixes a bug when all the dependent variables are the same. I'm not a mathmatician in any way, so if you have any better solutions I'd gladly hear them, but this (from a development point of view..) fixes the issue we had when all dependent values were the same. It would cause a division by zero.